Migrate dom funcs to its own file.
